requirements.txt
torch>=1.7.1
torchvision>=0.8.2
torchdiffeq
geotorch
requests>=2.25.1
numpy>=1.19.4
Jinja2~=3.1.2
timm>=0.9.0
tqdm>=4.56.1
pandas>=1.3.5
git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e#egg=autoattack
gdown==5.1.0
pyyaml
yacs==0.1.8